refactor: 重构 QueryGroup
